core/testing/requirements.txt
flask>=3.0
pandas>=2.0
openpyxl>=3.1
anthropic>=0.25
google-cloud-bigquery>=3.0
google-auth>=2.0
python-dotenv>=1.0
sqlalchemy>=2.0
pymysql>=1.1
pyyaml>=6.0
claude-agent-sdk
anyio>=4.0